Download command added to backend.
