+++ src-tauri/src/lib.rs
@@ -9,6 +9,7 @@ use std::{
 use serde::Serialize;
 use tauri::ipc::Channel;
 
+mod commands;
 mod ytdlp;
 
 #[tauri::command]
@@ -26,85 +27,89 @@ async fn fetch_data(url: &str) -> Result<ytdlp::YtdlpResponse, String> {
     }
 }
 
-#[derive(Clone, Serialize)]
-#[serde(
-    rename_all = "camelCase",
-    rename_all_fields = "camelCase",
-    tag = "event",
-    content = "data"
-)]
-enum DownloadEvent {
-    Started {
-        url: String,
-        audio_format: String,
-        video_format: String,
-        // download_id: usize,
-    },
-    Progress {
-        // download_id: usize,
-        output: String,
-    },
-    Finished {
-        // download_id: usize,
-    },
-}
+// #[derive(Clone, Serialize)]
+// #[serde(
+//     rename_all = "camelCase",
+//     rename_all_fields = "camelCase",
+//     tag = "event",
+//     content = "data"
+// )]
+// enum DownloadEvent {
+//     Started {
+//         url: String,
+//         audio_format: String,
+//         video_format: String,
+//         // download_id: usize,
+//     },
+//     Progress {
+//         // download_id: usize,
+//         output: String,
+//     },
+//     Finished {
+//         // download_id: usize,
+//     },
+// }
 
-#[tauri::command]
-async fn download(
-    url: &str,
-    audio_format: &str,
-    video_format: &str,
-    output_dir: &str,
-    on_event: Channel<DownloadEvent>,
-) -> Result<(), ()> {
-    // let output = Command::new("yt-dlp").arg(format!("-F {}", url)).output();
-    let mut cmd = Command::new("yt-dlp")
-        .arg(format!("-f {}+{}", audio_format, video_format))
-        .arg("--force-overwrites")
-        .arg("-o")
-        .arg(format!("{}\\%(title)s.%(ext)s", output_dir))
-        .arg("https://www.youtube.com/watch?v=Dl2vf04UCAM")
-        .stdout(Stdio::piped())
-        // .stderr(Stdio::piped())
-        .spawn()
-        .unwrap();
+// #[tauri::command]
+// async fn download(
+//     url: &str,
+//     audio_format: &str,
+//     video_format: &str,
+//     output_dir: &str,
+//     on_event: Channel<DownloadEvent>,
+// ) -> Result<(), ()> {
+//     // let output = Command::new("yt-dlp").arg(format!("-F {}", url)).output();
+//     let mut cmd = Command::new("yt-dlp")
+//         .arg(format!("-f {}+{}", audio_format, video_format))
+//         .arg("--force-overwrites")
+//         .arg("-o")
+//         .arg(format!("{}\\%(title)s.%(ext)s", output_dir))
+//         .arg("https://www.youtube.com/watch?v=Dl2vf04UCAM")
+//         .stdout(Stdio::piped())
+//         // .stderr(Stdio::piped())
+//         .spawn()
+//         .unwrap();
 
-    on_event.send(DownloadEvent::Started {
-        url: url.to_string(),
-        audio_format: audio_format.to_string(),
-        video_format: video_format.to_string(),
-    });
+//     on_event.send(DownloadEvent::Started {
+//         url: url.to_string(),
+//         audio_format: audio_format.to_string(),
+//         video_format: video_format.to_string(),
+//     });
 
-    {
-        let stdout = cmd.stdout.as_mut().unwrap();
-        let stdout_reader = BufReader::new(stdout);
+//     {
+//         let stdout = cmd.stdout.as_mut().unwrap();
+//         let stdout_reader = BufReader::new(stdout);
 
-        // Every line that displays download progress in output includes \r instead of \n. By
-        // splitting them we can parse out the progress %, download speed and file size.
-        let stdout_lines = stdout_reader.split(b'\r');
+//         // Every line that displays download progress in output includes \r instead of \n. By
+//         // splitting them we can parse out the progress %, download speed and file size.
+//         let stdout_lines = stdout_reader.split(b'\r');
 
-        for line in stdout_lines {
-            let str = String::from_utf8(line.unwrap()).unwrap();
+//         for line in stdout_lines {
+//             let str = String::from_utf8(line.unwrap()).unwrap();
 
-            println!("{:?}", str);
+//             println!("{:?}", str);
 
-            on_event
-                .send(DownloadEvent::Progress { output: str })
-                .unwrap();
-        }
+//             on_event
+//                 .send(DownloadEvent::Progress { output: str })
+//                 .unwrap();
+//         }
 
-        on_event.send(DownloadEvent::Finished {}).unwrap();
-    }
+//         on_event.send(DownloadEvent::Finished {}).unwrap();
+//     }
 
-    Ok(())
-}
+//     Ok(())
+// }
 
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
     tauri::Builder::default()
         .plugin(tauri_plugin_dialog::init())
         .plugin(tauri_plugin_opener::init())
-        .invoke_handler(tauri::generate_handler![greet, fetch_data, download])
+        .invoke_handler(tauri::generate_handler![
+            greet,
+            fetch_data,
+            commands::download
+        ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
 }
